requirements.txt
Flask~=1.1.2
Flask-JWT-Extended~=3.13.1
Flask-RESTful
Flask-SQLAlchemy
bcrypt~=3.2.0
pytz~=2021.1
datetime~=4.3
tzlocal~=2.1
uwsgi
psycopg2